refactor: 이미지업로드 컴포넌트 처리
