+++ src/pages/ProjectCreationPage.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import styled from "styled-components";
 import Button from "../components/Button";
 import TagInput from "../components/TagInputContainer";
-import NavigationBar from "../components/NavigationBar";
+import ImageUpload from "../components/ImageUpload";
 import { useDispatch, useSelector } from "react-redux";
 import {
   setProjectData,
@@ -39,17 +39,8 @@ const ProjectCreationPage = () => {
     dispatch(setProjectData({ ...projectState, [field]: value }));
   };
 
-  // 이미지 업로드 핸들러
-  const handleImageUpload = (e) => {
-    const file = e.target.files[0];
-    if (file) {
-      const reader = new FileReader();
-      reader.onloadend = () => {
-        setImagePreview(reader.result);
-        handleInputChange("image", reader.result);
-      };
-      reader.readAsDataURL(file);
-    }
+  const handleImageChange = (newImage) => {
+    handleInputChange("image", newImage);
   };
 
   // 멤버 인원수 변경 핸들러
@@ -150,26 +141,12 @@ const ProjectCreationPage = () => {
 
         <Label>
           대표 이미지
-          <ImageWrapper>
-            <ImagePreview
-              src={imagePreview || "src/assets/icon/defaultImg.png"}
-              alt="미리보기"
-            />
-            <UploadButton>
-              <button>
-                <label htmlFor="image-upload">이미지 업로드</label>
-              </button>
-              <input
-                type="file"
-                accept="image/*"
-                onChange={handleImageUpload}
-                id="image-upload"
-              />
-            </UploadButton>
-          </ImageWrapper>
+          <ImageUpload
+            initialImage={projectState.image}
+            onImageChange={handleImageChange}
+          />
           {errors.image && <ErrorText>{errors.image}</ErrorText>}
         </Label>
-
 
         <Label>모집 인원
           {(projectState.platform_dto_list || []).map((member, index) => (
@@ -328,39 +305,6 @@ const MemberWrapper = styled.div`
   gap: 10px;
   margin-bottom: 10px;
 `;
-
-const ImageWrapper = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-const ImagePreview = styled.img`
-  width: 300px;
-  height: 190px;
-  margin: 8px 0;
-  border-radius: 10px;
-  object-fit: cover;
-  background-color: lightgray; 
-`;
-
-const UploadButton = styled.div`
-  margin-left: 20px;
-
-  button {
-    cursor: pointer;
-    background-color: white;
-    color: ${(props) => props.theme.colors.accent};
-    padding: 10px;
-    border-radius: 5px;
-    font-size: 14px;
-    border: 1px solid ${(props) => props.theme.colors.accent};
-  }
-
-  input[type="file"] {
-    display: none;
-  }
-`;
-
 
 const ButtonWrapper = styled.div`
   margin-top: 20px;
